+++ src/AnalysisG/modules/nusol/conuix/include/conuix/matrix.h
@@ -8,20 +8,20 @@ struct matrix_t {
         matrix_t dot(const matrix_t& o);
         matrix_t T() const;
        
-        double& at(int _r, int _c);
-        const double& at(int _r, int _c) const;
+        long double& at(int _r, int _c);
+        const long double& at(int _r, int _c) const;
 
         matrix_t& operator=(const matrix_t& other);
         matrix_t& operator+(const matrix_t& o) const;
         matrix_t& operator-(const matrix_t& o) const;
-        matrix_t& operator*(double s) const;
+        matrix_t& operator*(long double s) const;
 
         void print(int p = 9);
 
     private:
         int r = -1; 
         int c = -1;
-        double** data = nullptr; 
+        long double** data = nullptr; 
 };
