+++ src/AnalysisG/modules/nusol/conuix/include/conuix/memory.h
@@ -1,33 +1,33 @@
 #ifndef H_CONUIX_MEMORY
 #define H_CONUIX_MEMORY
 
-double** alloc(
+long double** alloc(
     const int dim_i, const int dim_j
 ); 
 
-double** flush(
-    double** dmx, const int dim_j
+long double** flush(
+    long double** dmx, const int dim_j
 ); 
 
 void copy(
-    double** from, double** to, 
+    long double** from, long double** to, 
     const int dim_i, const int dim_j
 ); 
 
 void ops(
-    double** O, double** A, double** B, 
+    long double** O, long double** A, long double** B, 
     const int dim_i, const int dim_j, 
     const float sign
 ); 
 
 void opm(
-    double** O, double** A, 
+    long double** O, long double** A, 
     const int dim_i, const int dim_j, 
     const float c
 ); 
 
 void opt(
-    double** O, double** A, 
+    long double** O, long double** A, 
     const int dim_i, const int dim_j
 );
